backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
pydantic==2.9.2
pydantic-settings==2.5.2
python-multipart==0.0.9
numpy==1.26.4
Pillow==10.4.0
faiss-cpu==1.8.0.post1
opencv-python==4.10.0.84
# Optional heavy deps (install if use_mock_models=False)
# sentence-transformers==3.1.1
# torch==2.4.1
# torchvision==0.19.1
# openai-whisper==20240930
